:sparkles: feat: add use case for fetch accounts with portfolios, adjust some payloads and UI
