--- src/app/(routes)/ledgers/[id]/accounts-and-portfolios/accounts-portfolios-tab-content.tsx
@@ -1,11 +1,0 @@
-import { AccountsContent } from './accounts-content'
-import { PortfoliosContent } from './portfolios-content'
-
-export const AccountsPortfoliosTabContent = () => {
-  return (
-    <div className="flex flex-1 flex-col gap-6">
-      <PortfoliosContent />
-      <AccountsContent />
-    </div>
-  )
-}
