+++ src/app/(routes)/ledgers/[id]/accounts/accounts-tab-content.tsx
@@ -1,0 +1,5 @@
+import { AccountsContent } from './accounts-content'
+
+export const AccountsTabContent = () => {
+  return <AccountsContent />
+}
