+++ src/app/(routes)/ledgers/[id]/portfolios/portfolios-tab-content.tsx
@@ -1,0 +1,9 @@
+import { PortfoliosContent } from './portfolios-content'
+
+export const PortfoliosTabContent = () => {
+  return (
+    <div className="flex flex-1 flex-col">
+      <PortfoliosContent />
+    </div>
+  )
+}
