+++ src/client/accounts.ts
@@ -17,19 +17,17 @@ import {
 type UseListAccountsProps = {
   organizationId: string
   ledgerId: string
-  portfolioId: string
 }
 
 export const useListAccounts = ({
   organizationId,
   ledgerId,
-  portfolioId,
   ...options
 }: UseListAccountsProps) => {
   return useQuery<PaginationDto<AccountEntity>>({
-    queryKey: [organizationId, ledgerId, portfolioId, 'accounts'],
+    queryKey: [organizationId, ledgerId, 'accounts'],
     queryFn: getFetcher(
-      `/api/organizations/${organizationId}/ledgers/${ledgerId}/portfolios/${portfolioId}/accounts`
+      `/api/organizations/${organizationId}/ledgers/${ledgerId}/accounts`
     ),
     ...options
   })
@@ -40,15 +38,15 @@ type UseAllPortfoliosAccountsProps = {
   ledgerId: string
 }
 
-export const useAllPortfoliosAccounts = ({
+export const useAllAccountsWithPortfolios = ({
   organizationId,
   ledgerId,
   ...options
 }: UseAllPortfoliosAccountsProps) => {
   return useQuery<PaginationDto<PortfolioViewResponseDTO>>({
-    queryKey: [organizationId, ledgerId, 'portfolios-accounts'],
+    queryKey: [organizationId, ledgerId, 'accounts-with-portfolios'],
     queryFn: getFetcher(
-      `/api/organizations/${organizationId}/ledgers/${ledgerId}/portfolios/portfolios-accounts`
+      `/api/organizations/${organizationId}/ledgers/${ledgerId}/accounts-portfolios`
     ),
     ...options
   })
@@ -57,21 +55,19 @@ export const useAllPortfoliosAccounts = ({
 type UseDeleteAccountProps = UseMutationOptions & {
   organizationId: string
   ledgerId: string
-  portfolioId: string
   accountId: string
 }
 
 export const useDeleteAccount = ({
   organizationId,
   ledgerId,
-  portfolioId,
   accountId,
   ...options
 }: UseDeleteAccountProps) => {
   return useMutation<any, any, any>({
-    mutationKey: [organizationId, ledgerId, portfolioId, accountId],
+    mutationKey: [organizationId, ledgerId, accountId],
     mutationFn: deleteFetcher(
-      `/api/organizations/${organizationId}/ledgers/${ledgerId}/portfolios/${portfolioId}/accounts`
+      `/api/organizations/${organizationId}/ledgers/${ledgerId}/accounts`
     ),
     ...options
   })
@@ -80,19 +76,17 @@ export const useDeleteAccount = ({
 type UseCreateAccountProps = UseMutationOptions & {
   organizationId: string
   ledgerId: string
-  portfolioId: string
 }
 
 export const useCreateAccount = ({
   organizationId,
   ledgerId,
-  portfolioId,
   ...options
 }: UseCreateAccountProps) => {
   return useMutation<any, any, any>({
-    mutationKey: [organizationId, ledgerId, portfolioId],
+    mutationKey: [organizationId, ledgerId],
     mutationFn: postFetcher(
-      `/api/organizations/${organizationId}/ledgers/${ledgerId}/portfolios/${portfolioId}/accounts`
+      `/api/organizations/${organizationId}/ledgers/${ledgerId}/accounts`
     ),
     ...options
   })
@@ -101,21 +95,19 @@ export const useCreateAccount = ({
 type UseUpdateAccountProps = UseMutationOptions & {
   organizationId: string
   ledgerId: string
-  portfolioId: string
   accountId: string
 }
 
 export const useUpdateAccount = ({
   organizationId,
   ledgerId,
-  portfolioId,
   accountId,
   ...options
 }: UseUpdateAccountProps) => {
   return useMutation<any, any, any>({
-    mutationKey: [organizationId, ledgerId, portfolioId, accountId],
+    mutationKey: [organizationId, ledgerId, accountId],
     mutationFn: patchFetcher(
-      `/api/organizations/${organizationId}/ledgers/${ledgerId}/portfolios/${portfolioId}/accounts/${accountId}`
+      `/api/organizations/${organizationId}/ledgers/${ledgerId}/accounts/${accountId}`
     ),
     ...options
   })
