+++ src/core/application/use-cases/accounts/create-account-use-case.ts
@@ -8,7 +8,6 @@ export interface CreateAccount {
   execute: (
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     account: CreateAccountDto
   ) => Promise<AccountResponseDto>
 }
@@ -23,7 +22,6 @@ export class CreateAccountUseCase implements CreateAccount {
   async execute(
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     account: CreateAccountDto
   ): Promise<AccountResponseDto> {
     account.status = {
@@ -31,10 +29,10 @@ export class CreateAccountUseCase implements CreateAccount {
       description: 'Active Account'
     }
     const accountEntity: AccountEntity = AccountMapper.toDomain(account)
+    console.log('accountEntity', accountEntity)
     const accountCreated = await this.createAccountRepository.create(
       organizationId,
       ledgerId,
-      portfolioId,
       accountEntity
     )
