+++ src/core/domain/repositories/accounts/create-accounts-repository.ts
@@ -4,7 +4,6 @@ export abstract class CreateAccountsRepository {
   abstract create: (
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     account: AccountEntity
   ) => Promise<AccountEntity>
 }
