+++ src/core/domain/repositories/accounts/delete-accounts-repository.ts
@@ -2,7 +2,6 @@ export abstract class DeleteAccountsRepository {
   abstract delete: (
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     accountId: string
   ) => Promise<void>
 }
