+++ src/core/domain/repositories/accounts/fetch-account-by-id-repository.ts
@@ -4,7 +4,6 @@ export abstract class FetchAccountByIdRepository {
   abstract fetchById: (
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     accountId: string
   ) => Promise<AccountEntity>
 }
