+++ src/core/domain/repositories/accounts/fetch-all-accounts-repository.ts
@@ -5,7 +5,6 @@ export abstract class FetchAllAccountsRepository {
   abstract fetchAll: (
     organizationId: string,
     ledgerId: string,
-    portfolioId: string,
     limit: number,
     page: number
   ) => Promise<PaginationEntity<AccountEntity>>
