+++ src/schema/account.ts
@@ -3,15 +3,15 @@ import { metadata } from './metadata'
 
 const name = z.string().min(3).max(255)
 
-const alias = z.string().min(1).max(255).optional()
+const alias = z.string().min(1).max(255)
 
-const entityId = z.string().min(1).max(255).optional()
+const entityId = z.string().nullable().optional()
 
 const assetCode = z.string()
 
-const portfolioId = z.string()
+const portfolioId = z.string().nullable().optional()
 
-const productId = z.string()
+const productId = z.string().nullable().optional()
 
 const type = z.string()
