+++ src/types/accounts-type.ts
@@ -4,18 +4,21 @@ export type AccountType = {
   assetCode: string
   organizationId: string
   name: string
-  alias: string
-  type: string
-  entityId: string
+  alias?: string
+  type?: string
+  entityId?: string
   parentAccountId: string
-  portfolioId: string
+  portfolioId?: string | null
+  portfolio: {
+    name: string
+  }
   portfolioName?: string
   productId: string
   status: {
     code: string
     description: string
   }
-  metadata: Record<string, any>
+  metadata?: Record<string, any>
   createdAt: Date
   updatedAt: Date
   deletedAt: Date | null
